Fix failing tests Lint with ruff, black, isort, flake8
