+++ pyproject.toml
@@ -17,6 +17,22 @@ mysqlclient = "^2.2.0"
 fs-s3fs = { version = "^1.1.1", optional = true }
 sshtunnel = "0.4.0"
 
+[tool.poetry.dev-dependencies]
+pytest = "^7.3.2"
+tox = "^3.24.4"
+flake8 = "^5.0.4"
+flake8-docstrings = "^1.7.0"
+black = "23.1.0"
+mypy = "^1.0"
+isort = "^5.11.5"
+remote-pdb="2.1.0"
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3 # Vertical Hanging Indent
+src_paths = "tap_mysql"
+
+
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.1"
 faker = ">=18.5.1,<20.0.0"
