+++ tests/test_ssh_tunnel.py
@@ -15,7 +15,8 @@ SAMPLE_CONFIG = {
     },
 }
 
-def test_ssh_tunnel():
-    """We expect the SSH environment to already be up"""
+
+def test_ssh_tunnel() -> None:
+    """We expect the SSH environment to already be up."""
     tap = TapMySQL(config=SAMPLE_CONFIG)
     tap.sync_all()
