+++ tox.ini
@@ -1,19 +1,48 @@
 # This file can be used to customize tox tests as well as other test frameworks like flake8 and mypy
 
 [tox]
-envlist = py37, py38, py39, py310, py311
+envlist = py39
+; envlist = py37, py38, py39
 isolated_build = true
 
 [testenv]
-allowlist_externals = poetry
+whitelist_externals = poetry
+
 commands =
     poetry install -v
     poetry run pytest
+    poetry run black --check tap_mysql
+    poetry run flake8  tap_mysql
 
 [testenv:pytest]
 # Run the python tests.
 # To execute, run `tox -e pytest`
-envlist = py37, py38, py39, py310, py311
+envlist = py37, py38, py39
 commands =
     poetry install -v
     poetry run pytest
+
+[testenv:format]
+# Attempt to auto-resolve lint errors before they are raised.
+# To execute, run `tox -e format`
+commands =
+    poetry install -v
+    poetry run black tap_mysql
+    poetry run isort tap_mysql
+
+[testenv:lint]
+# Raise an error if lint and style standards are not met.
+# To execute, run `tox -e lint`
+commands =
+    poetry install -v
+    poetry run black --check --diff tap_mysql/
+    poetry run isort --check tap_mysql
+    poetry run flake8 tap_mysql
+
+[flake8]
+ignore = W503
+max-line-length = 88
+max-complexity = 10
+
+[pydocstyle]
+ignore = D105,D203,D213
